added homepage, added react router, also organized folders a bit, please install react router
