+++ bigbyte/src/App.js
@@ -1,5 +1,8 @@
 import './App.css';
-import Login from './Login.js';
+import Login from './pages/Login.js';
+import Home from './pages/Home.js';
+import ReactDOM from "react-dom/client";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 function App() {
   return (
@@ -20,10 +23,18 @@ function App() {
     //     </a>
     //   </header>
     // </div>
-    <div className="Login">
-      <Login />
+    <BrowserRouter>
+    <Routes>
+      
+        <Route index element={<Home />} />
+        <Route path="Login" element= {<Login /> }/>
+        
+     
+    </Routes>
+  </BrowserRouter>
 
-    </div>
+    
+    
   );
 }
